Add boundVOI codeboxes
One codebox demonstrates that boundVOI acts optimally in bandit POMDPs
when there are only two arms. The other two show the scaling of the
runtime of the boundVOI agent. I have also edited the codebox showing
the scaling of the runtime of the optimal agent in a bandit POMDP as the
lifetime varies, changing the output plot to a line graph.