Display trajectories of discounters in gridworld
This commit displays the trajectories of the naive and sophisticated
hyperbolic and exponential discounters in the restaurant choice
gridworld. For hyperbolic discounters, naive vs sophisticated makes a
difference, for exponential discounters it doesn't. Trajectories are now
nicely displayed, and timeLeft was changed to stop the agent from going
to Donut North the long way in order to only recieve the first (positive)
reward there.